separate query-sending code from response-receiving code
